Updated HTTP status code descriptions to match RFC 9110.
This patch changes the status codes and status pages served by
freenginx to use the wording from RFC 9110.

Exception for HTTP status code 413 "Request Entity Too Large",
which is "Content Too Large" in the RFC. In that case we would
lose information about what actually is too large, the request
or the response.